test: organize Miri coverage by unsafe boundary
Preserve all 181 lang and SPL cases while replacing two monolithic integration files with named ownership modules.
